json_data_gen/requirements.txt
Faker
Flask
google-cloud-storage
google-cloud-pubsub